Se hacen varios cambios durante el proceso de documentación. Se le cambia el nombre a un par de clases, y se hacen algunos refactor pequeños.

+++ src/common/Circle.h
@@ -7,22 +7,22 @@ struct Circle {
     int x, y;
     int r;
 
-    //FIXME: duplicate of Utils::distanceSquared
     double distanceSquared(int x1, int y1, int x2, int y2) {
-        int deltaX = x2 - x1; int deltaY = y2 - y1;
+        int deltaX = x2 - x1;
+        int deltaY = y2 - y1;
         return deltaX * deltaX + deltaY * deltaY;
     }
 
-    /* Circles overlap iff distance between centers is less than sum of radii
-     *
-     * Due to the monotonicity of polynomials, comparing the squares of
-     * these quantities is equivalent to comparing the quantities themselves.
-     * (proof of this theorem sold separately)
+    /* Los círculos se superponen si la distancia entre centros es menor que
+     * la suma de radios.
+     *
+     * Debido a la monotonicidad de los polinomios, comparando los cuadrados de
+     * estas cantidades equivalen a comparar las cantidades mismas.
+     * (prueba de este teorema se vende por separado).
      */
     bool hasCollisionWith(Circle &b){
         /* Square of the sum of the radii */
-        int r2 = std::pow(r + b.r, 2);
-
+        double r2 = std::pow(r + b.r, 2);
         return (distanceSquared(x, y, b.x, b.y) < r2);
     }
 };
